Move PgRepositories to new files
