+++ src/components/authentication/authentication.module.ts
@@ -3,10 +3,8 @@ import { APP_INTERCEPTOR } from '@nestjs/core';
 import { splitDb } from '../../core';
 import { AuthorizationModule } from '../authorization/authorization.module';
 import { UserModule } from '../user/user.module';
-import {
-  AuthenticationRepository,
-  PgAuthenticationRepository,
-} from './authentication.repository';
+import { PgAuthenticationRepository } from './authentication.pg.repository';
+import { AuthenticationRepository } from './authentication.repository';
 import { AuthenticationService } from './authentication.service';
 import { CryptoService } from './crypto.service';
 import { LoginResolver } from './login.resolver';
@@ -32,7 +30,6 @@ import { SessionResolver } from './session.resolver';
     SessionInterceptor,
     { provide: APP_INTERCEPTOR, useExisting: SessionInterceptor },
     splitDb(AuthenticationRepository, PgAuthenticationRepository),
-    PgAuthenticationRepository,
   ],
   exports: [AuthenticationService, CryptoService, AuthenticationRepository],
 })
