+++ src/components/field-zone/field-zone.module.ts
@@ -3,10 +3,8 @@ import { splitDb } from '../../core';
 import { AuthorizationModule } from '../authorization/authorization.module';
 import { UserModule } from '../user/user.module';
 import { FieldZoneLoader } from './field-zone.loader';
-import {
-  FieldZoneRepository,
-  PgFieldZoneRepository,
-} from './field-zone.repository';
+import { PgFieldZoneRepository } from './field-zone.pg.repository';
+import { FieldZoneRepository } from './field-zone.repository';
 import { FieldZoneResolver } from './field-zone.resolver';
 import { FieldZoneService } from './field-zone.service';
 
@@ -21,7 +19,6 @@ import { FieldZoneService } from './field-zone.service';
     FieldZoneRepository,
     FieldZoneLoader,
     splitDb(FieldZoneRepository, PgFieldZoneRepository),
-    PgFieldZoneRepository,
   ],
   exports: [FieldZoneService],
 })
