+++ src/components/funding-account/funding-account.module.ts
@@ -2,10 +2,8 @@ import { forwardRef, Module } from '@nestjs/common';
 import { splitDb } from '../../core';
 import { AuthorizationModule } from '../authorization/authorization.module';
 import { FundingAccountLoader } from './funding-account.loader';
-import {
-  FundingAccountRepository,
-  PgFundingAccountRepository,
-} from './funding-account.repository';
+import { PgFundingAccountRepository } from './funding-account.pg.repository';
+import { FundingAccountRepository } from './funding-account.repository';
 import { FundingAccountResolver } from './funding-account.resolver';
 import { FundingAccountService } from './funding-account.service';
 
@@ -17,7 +15,6 @@ import { FundingAccountService } from './funding-account.service';
     FundingAccountRepository,
     FundingAccountLoader,
     splitDb(FundingAccountRepository, PgFundingAccountRepository),
-    PgFundingAccountRepository,
   ],
   exports: [FundingAccountService],
 })
